requirement.txt
Django==2.2.9
django-cors-headers==3.2.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
mysqlclient==1.4.6
python-decouple==3.3
django-simple-history==2.8.0
numpy==1.18.2
pandas==1.0.3
python-dateutil==2.8.1